pulling apart the config file a bit
