+++ packages/cli-lib/lib/__tests__/config.js
@@ -10,7 +10,6 @@ const {
   updateAccountConfig,
   validateConfig,
   deleteEmptyConfigFile,
-  configFilenameIsIgnoredByGitignore,
   setConfigPath,
   createEmptyConfigFile,
 } = require('../config');
@@ -525,98 +524,11 @@ describe('lib/config', () => {
     });
   });
 
-  describe('configFilenameIsIgnoredByGitignore method', () => {
-    const fs = require('fs-extra');
-
-    afterAll(() => {
-      setConfigPath(CONFIG_PATHS.default);
-      mockedConfigPath = CONFIG_PATHS.default;
-    });
-
-    it('returns false if the config file is not ignored', () => {
-      const gitignoreContent = '';
-      setConfigPath(`/Users/fakeuser/someproject/hubspot.config.yml`);
-      jest.spyOn(fs, 'readFileSync').mockImplementation(() => {
-        return Buffer.from(gitignoreContent);
-      });
-
-      expect(
-        configFilenameIsIgnoredByGitignore([
-          '/Users/fakeuser/someproject/.gitignore',
-        ])
-      ).toBe(false);
-    });
-
-    it('identifies if a config file is ignored with a specific ignore statement', () => {
-      const gitignoreContent = 'hubspot.config.yml';
-      setConfigPath(`/Users/fakeuser/someproject/hubspot.config.yml`);
-      const readFileSyncSpy = jest
-        .spyOn(fs, 'readFileSync')
-        .mockImplementation(() => {
-          return Buffer.from(gitignoreContent);
-        });
-
-      expect(
-        configFilenameIsIgnoredByGitignore([
-          '/Users/fakeuser/someproject/.gitignore',
-        ])
-      ).toBe(true);
-      readFileSyncSpy.mockReset();
-    });
-
-    it('identifies if a config file is ignored with a wildcard statement', () => {
-      const gitignoreContent = 'hubspot.config.*';
-      setConfigPath(`/Users/fakeuser/someproject/hubspot.config.yml`);
-      const readFileSyncSpy = jest
-        .spyOn(fs, 'readFileSync')
-        .mockImplementation(() => {
-          return Buffer.from(gitignoreContent);
-        });
-
-      expect(
-        configFilenameIsIgnoredByGitignore([
-          '/Users/fakeuser/someproject/.gitignore',
-        ])
-      ).toBe(true);
-      readFileSyncSpy.mockReset();
-    });
-
-    it('identifies if a non-standard named config file is not ignored', () => {
-      const gitignoreContent = 'hubspot.config.yml';
-      setConfigPath('/Users/fakeuser/someproject/config/my.custom.name.yml');
-      const readFileSyncSpy = jest
-        .spyOn(fs, 'readFileSync')
-        .mockImplementation(() => {
-          return Buffer.from(gitignoreContent);
-        });
-
-      expect(
-        configFilenameIsIgnoredByGitignore([
-          '/Users/fakeuser/someproject/.gitignore',
-        ])
-      ).toBe(false);
-      readFileSyncSpy.mockReset();
-    });
-
-    it('identifies if a non-standard named config file is ignored', () => {
-      const gitignoreContent = 'my.custom.name.yml';
-      setConfigPath('/Users/fakeuser/someproject/config/my.custom.name.yml');
-      const readFileSyncSpy = jest
-        .spyOn(fs, 'readFileSync')
-        .mockImplementation(() => {
-          return Buffer.from(gitignoreContent);
-        });
-
-      expect(
-        configFilenameIsIgnoredByGitignore([
-          '/Users/fakeuser/someproject/.gitignore',
-        ])
-      ).toBe(true);
-      readFileSyncSpy.mockReset();
-    });
-  });
-
   describe('getConfigPath method', () => {
+    beforeAll(() => {
+      setConfigPath(CONFIG_PATHS.default);
+    });
+
     describe('when a standard config is present', () => {
       it('returns the standard config path', () => {
         const configPath = getConfigPath();
